+++ sdk/containerregistry/azure-containerregistry/tests/test_container_registry_client_async.py
@@ -233,24 +233,24 @@ class TestContainerRegistryClient(AsyncContainerRegistryTestClass):
         assert received.can_list == True
 
     @acr_preparer()
-    async def test_get_tag_properties(self, containerregistry_endpoint):
+    async def test_get_tag(self, containerregistry_endpoint):
         repo = self.get_resource_name("repo")
         tag = self.get_resource_name("tag")
         self.import_image(HELLO_WORLD, ["{}:{}".format(repo, tag)])
 
         client = self.create_registry_client(containerregistry_endpoint)
 
-        properties = await client.get_tag_properties(repo, tag)
+        properties = await client.get_tag(repo, tag)
 
         assert isinstance(properties, ArtifactTagProperties)
         assert properties.name == tag
 
     @acr_preparer()
-    async def test_get_tag_properties_does_not_exist(self, containerregistry_endpoint):
+    async def test_get_tag_does_not_exist(self, containerregistry_endpoint):
         client = self.create_registry_client(containerregistry_endpoint)
 
         with pytest.raises(ResourceNotFoundError):
-            await client.get_tag_properties("Nonexistent", "Nonexistent")
+            await client.get_tag("Nonexistent", "Nonexistent")
 
     @acr_preparer()
     async def test_set_tag_properties(self, containerregistry_endpoint):
@@ -260,7 +260,7 @@ class TestContainerRegistryClient(AsyncContainerRegistryTestClass):
 
         client = self.create_registry_client(containerregistry_endpoint)
 
-        properties = await client.get_tag_properties(repo, tag)
+        properties = await client.get_tag(repo, tag)
         properties.can_delete = False
         properties.can_read = False
         properties.can_write = False