+++ sdk/containerregistry/azure-containerregistry/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Removes `DeleteRepositoryResult`. `ContainerRegistryClient.delete_repository` now returns `None`
 * Removed `writeable_properties` objects, placing `can_delete/read/write/list` properties on the immediate `Repository/Tag/ArtifactManifestProperties` objects
 * Removed `ContainerRepository` and `RegistryArtifact` classes. The methods for acting on a repository and a registry artifact are now contained in the `ContainerRegistryClient` object.
+* Renamed `ContainerRegistryClient.get_tag_properties` to `ContainerRegistryClient.get_tag`
 
 ## 1.0.0b2 (2021-05-11)
 * Rename `DeletedRepositoryResult` to `DeleteRepositoryResult`